Fixed issues with user creation
There were a couple of bugs with user creation, one to do with the kdf argument being wrong and a couple
to do with failing to pass the enum properly. For future reference, you need to pass your Python enum to
the sqlalchemy enum as an argument in the ORM class definition.